% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformalInference.multi-package.R
\docType{package}
\name{conformalInference.multi-package}
\alias{conformalInference.multi}
\alias{conformalInference.multi-package}
\title{conformalInference.multi: Conformal Inference Tools for Regression with Multivariate Response}
\description{
It computes full conformal, split conformal and multi split conformal prediction regions when the response variable is multivariate (i.e. dimension is greater than one). Moreover, the package also contain plot functions to visualize the output of the full and split conformal functions. To guarantee consistency, the package structure mimics the univariate conformalInference package of professor Ryan Tibshirani.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ryantibs/conformal}
}

}
\author{
\strong{Maintainer}: Paolo Vergottini \email{paolo.vergottini@gmail.com}

Authors:
\itemize{
  \item Jacopo Diquigiovanni [thesis advisor]
  \item Matteo Fontana \email{matteo.fontana@ec.europa.eu} [thesis advisor]
  \item Aldo Solari [thesis advisor]
  \item Simone Vantini [thesis advisor]
}

Other contributors:
\itemize{
  \item Ryan Tibshirani [contributor]
}

}
\keyword{internal}
