% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.full.R
\name{plot_multidim_full}
\alias{plot_multidim_full}
\title{Plot Confidence Regions obtained from Full Conformal}
\usage{
plot_multidim_full(full)
}
\arguments{
\item{full}{It's the output of the multivariate full conformal prediction function}
}
\value{
A list of ggplots (output[[i]] is the i-th observation confidence region).
}
\description{
Plot Confidence Regions obtained from Full Conformal
}
\details{
It exploits the package \code{\link{ggplot2}} and \code{\link{hrbrthemes}}
to better visualize the results.
}
\examples{
n=25
p=4
q=2


mu=rep(0,p)
x = mvtnorm::rmvnorm(n, mu)
beta<-sapply(1:q, function(k) c(mvtnorm::rmvnorm(1,mu)))
y = x\%*\%beta + t(mvtnorm::rmvnorm(q,1:n))
x0=x[ceiling(0.9*n):n,]
y0=y[ceiling(0.9*n):n,]

n0<-nrow(y0)
q<-ncol(y)


fun=mean_multi()

#################################### FULL CONFORMAL

final.full=conformal.multidim.full(x, y, x0, fun$train.fun,
                                fun$predict.fun, score="l2",
                                num.grid.pts.dim=5, grid.factor=1.25,
                                verbose=FALSE)

ppp<-plot_multidim_full(final.full)

}
